Move JsonFile data source code into data/json.go
